+++ commerce-api/src/main/java/pro/developia/api/dto/request/order/OrderCreateRequest.java
@@ -2,10 +2,12 @@ package pro.developia.api.dto.request.order;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class OrderCreateRequest {
     private Long memberId;
     private Long itemId;
     private Integer count;
-    private Integer totalPrice;
+    private BigDecimal totalPrice;
 }
